codecrafters submit [skip ci]

+++ src/aof.rs
@@ -2,7 +2,7 @@ use std::path::Path;
 
 use tokio::{
     fs::{self, File},
-    io::AsyncWriteExt,
+    io::{AsyncReadExt, AsyncWriteExt},
 };
 
 use crate::Config;
@@ -42,14 +42,25 @@ impl Aof {
             // Create file
             let base_filename = format!("{}.1.incr.aof", self.appendfilename);
             let filename = format!("{}{}", dirname, base_filename);
-            let mut file = File::create(filename).await.unwrap();
+            if !Path::exists(&Path::new(&filename)) {
+                let mut file = File::create(filename).await.unwrap();
+            }
 
             // Create manifest file
             let mf_base_filename = format!("{}.manifest", self.appendfilename);
             let mf_filename = format!("{}{}", dirname, mf_base_filename);
-            let line = format!("file {} seq 1 type i", base_filename);
-            let mut mf_file = File::create(mf_filename).await.unwrap();
-            let _ = mf_file.write_all(line.as_bytes()).await.unwrap();
+            if !Path::exists(&Path::new(&mf_filename)) {
+                let line = format!("file {} seq 1 type i", base_filename);
+                let mut mf_file = File::create(mf_filename.clone()).await.unwrap();
+                let _ = mf_file.write_all(line.as_bytes()).await.unwrap();
+            }
+
+            // Read manifest file
+            let mut mf_file = File::open(&Path::new(&mf_filename)).await.unwrap();
+            let mut buffer = String::new();
+            let _ = mf_file.read_to_string(&mut buffer).await.unwrap();
+            let aof_filename = buffer.split(' ').nth(1).unwrap();
+            println!("[aof] read aof file from manifest: {}", aof_filename);
         }
     }
 }
